test: add attributes tests and separate attributes arbitraries
