+++ test/html/create-element/attributes/arbitraries.js
@@ -1,0 +1,9 @@
+import { fc } from 'ava-fast-check'
+import { attributeNameArb } from './names/arbitraries'
+import { attributeValueArb } from './values/arbitraries'
+
+export const attributeArb = () =>
+  fc.tuple(attributeNameArb(), attributeValueArb())
+
+export const attributesArb = () =>
+  fc.dictionary(attributeNameArb(), attributeValueArb())
