+++ test/html/create-element/attributes/names/arbitraries.js
@@ -1,0 +1,3 @@
+import { fc } from 'ava-fast-check'
+
+export const attributeNameArb = fc.string
