+++ test/html/create-element/attributes/values/arbitraries.js
@@ -1,0 +1,11 @@
+import { fc } from 'ava-fast-check'
+
+export const attributeValueArb = () =>
+  fc.oneof(
+    fc.constant(null),
+    fc.constant(undefined),
+    fc.string(),
+    fc.integer(),
+    fc.float(),
+    fc.boolean()
+  )
